[0563] Removed unnecessary route scope - as it conflicts with courses itself - as it conflicts with course confirmation
Due to that large scale moving of files to its respective file locations
