+++ app/views/publish/courses/_basic_details_tab.html.erb
