+++ app/views/publish/courses/_course_abilities_blurb.html.erb
