+++ app/views/publish/courses/_course_table.html.erb
