+++ app/views/publish/courses/_course_table_row.html.erb
