+++ app/views/publish/courses/_description_content.html.erb
