+++ app/views/publish/courses/confirmation.html.erb
