+++ app/views/publish/courses/details.html.erb
