+++ app/views/publish/courses/index.html.erb
