+++ app/views/publish/courses/show.html.erb
