+++ app/views/publish/courses/vacancies/_multiple_sites.html.erb
