+++ app/views/publish/courses/vacancies/_single_site.html.erb
