+++ app/views/publish/courses/vacancies/_site_status_checkbox.html.erb
