+++ app/views/publish/courses/vacancies/edit.html.erb
